+++ src/painting/BorderRadiusGeometry.ts
@@ -1,0 +1,23 @@
+import { _Radius } from "../ui/Radius";
+
+class BorderRadiusGeometry {
+
+    topLeft?:_Radius;
+    topRight?:_Radius;
+    bottomLeft?:_Radius;
+    bottomRight?:_Radius;
+    topStart?:_Radius;
+    topEnd?:_Radius;
+    bottomStart?:_Radius;
+    bottomEnd?:_Radius;
+
+    constructor(args:{}) {
+        
+    }
+}
+
+
+
+export{
+    BorderRadiusGeometry
+}