+++ src/rendering/flex.ts
@@ -66,9 +66,6 @@ enum MainAxisAlignment {
     /** Place the children as close to the middle of the main axis as possible.*/
     center = "items-center",
   
-    /** Place the free space evenly between the children.*/
-    spaceBetween = "items-",
-  
     /** Place the free space evenly between the children as well as half of that
     * space before and after the first and last child.
     * */
@@ -88,7 +85,7 @@ class Flex{
     }
 }
 
-  export {
-      MainAxisAlignment,
-      CrossAxisAlignment
-  }
+export {
+    MainAxisAlignment,
+    CrossAxisAlignment
+}