+++ src/widgets/Button.ts
@@ -1,0 +1,10 @@
+import { _Container } from "./container.js";
+
+class Button extends _Container {
+    constructor(args:{}) {
+        
+        super({});
+        
+        return this;
+    }
+}