+++ src/widgets/framework.ts
@@ -78,7 +78,7 @@ export default class Widget {
 
     if (isNumber(args.width!)) {
       this.tag!.style.width = args.width!.toString() + this.width_size_measurement_unit;
-    }
+    }else{}
     if (isNumber(args.height!)) {
       this.tag!.style.height = args.height!.toString() + this.height_size_measurement_unit;
     }