+++ src/widgets/index.ts
@@ -1,11 +1,12 @@
 import { App } from "./app.js";
 import { Column } from "./column.js";
 import { Container } from "./container.js";
-import { Expanded } from "./exapnded.js";
+import { Expanded } from "./expanded.js";
 import { Widget } from "./framework.js";
 import { IconThemeData } from "./icon_theme_data.js";
 import { Padding } from "./padding.js";
 import { Row } from "./row.js";
+import { Spacer } from "./spacer.js";
 import { Text } from "./Text.js";
 
 export{
@@ -18,4 +19,5 @@ export{
     Padding,
     Widget,
     IconThemeData,
+    Spacer
 }